+++ src/logicblocks/event/persistence/in_memory/converters/__init__.py
