+++ src/logicblocks/event/persistence/in_memory/converters/clause.py
