+++ src/logicblocks/event/persistence/in_memory/converters/query.py
